atmel_lcdfb: introduce driver data
Introduce driver data like known from the kernel.
This allows us to get rid of the hack where the
intensity bit support was included in the lcd wiring mode.
(No longer any support for IBGR, IRBG)

It has the nice side-effect that all places where we test
for hacks can now use flags and not a set of cpu's.
So we keep all the configuration in one place.

The configuration is included for non-DT users.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>