arm: at91: refactor lowlevel_init selection
The aim with this conversion is to drop the dependency
where the mach-at91/ code call code in the boards file.
This dependency must be dropepd to enable multi-image builds.

The selection of lowlevel_init functions are pushed to
the individual boards.

Use AT91SAM926X_BOARD_INIT for the boards that share the
at91sam926x_board_init file, and push this to the
individual boards. To be used in follow-up patches

bootstrap is likewise pushed out to the individual boards.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Reviewed-by: Andrey Smirnov <andrew.smirnov@gmail.com>